requirements.txt
flask
Jinja2==3.0.1
joblib==1.0.1
gunicorn==20.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
pandas
numpy

Pillow==8.2.0
pipreqs==0.4.10
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
opencv-python-headless

six==1.16.0

threadpoolctl==2.1.0
urllib3==1.26.4
Werkzeug==2.0.1
yarg==0.1.9




keras
tensorflow-cpu
